+++ yolort/models/_utils.py
@@ -3,18 +3,6 @@ from typing import Tuple, Optional
 
 import torch
 from torch import nn, Tensor
-from torchvision.ops import box_iou
-
-
-def _evaluate_iou(target, pred):
-    """
-    Evaluate intersection over union (IOU) for target from dataset and
-    output prediction from model
-    """
-    if pred["boxes"].shape[0] == 0:
-        # no box detected, 0 IOU
-        return torch.tensor(0.0, device=pred["boxes"].device)
-    return box_iou(target["boxes"], pred["boxes"]).diag().mean()
 
 
 def _make_divisible(v: float, divisor: int, min_value: Optional[int] = None) -> int:
